feat(server): resolve libraries, movies, shows and episodes from the database
Adds mydia_api::mapping to turn mydia-db rows into the GraphQL types
Slice 1 declared, and wires it into the libraries, movie, tvShow and
episode resolvers. Widens MediaFile.size from i32 to i64 since a 4 GB
file overflows 32 bits; the SDL still says Int because async-graphql
renders every integer primitive that way, so sdl_parity stays green.

+++ server/crates/api/src/types/media.rs
@@ -85,7 +85,11 @@ pub struct MediaFile {
     pub codec: Option<String>,
     pub audio_codec: Option<String>,
     pub hdr_format: Option<String>,
-    pub size: Option<i32>,
+    /// Bytes. i64 rather than i32: a 4 GB film overflows 32 bits, and the
+    /// contract's `Int` is Absinthe's non-spec-compliant 2^53 Int, so the
+    /// Elixir server already emits real byte counts. async-graphql renders
+    /// i64 under the same `Int` scalar, so the SDL is unchanged.
+    pub size: Option<i64>,
     pub bitrate: Option<i32>,
     pub direct_play_supported: Option<bool>,
     pub stream_url: Option<String>,
